salary get on gui

+++ pharmacy_management_backend/src/main/java/com/omar/isdb62/pharmacy_management_backend/dto/RegisterRequest.java
@@ -19,6 +19,6 @@ public record RegisterRequest(
         String firstName,
         String lastName,
         String phoneNumber,
-        int salary
+        Long salary
 ) {
 }
