+++ pharmacy_management_backend/src/main/java/com/omar/isdb62/pharmacy_management_backend/dto/UserCreateRequest.java
@@ -10,6 +10,6 @@ public record UserCreateRequest(
         String firstName,
         String lastName,
         String phoneNumber,
-        int salary
+        Long salary
 ) {
 }
